+++ CHANGELOG.md
@@ -3,6 +3,7 @@ Yandex Turbo Pages Change Log
 
 Dev-master
 --------------------
+* Enh: added accordion method into Content helper
 * Enh: added rating method into Content helper
 
 1.1.1 July 10, 2018